refactor: split CKernel::run from kernel.cpp
